feat: Merge bingpai-related errors from `InvalidShoupaiError` into the `InvalidBingpai` variant
